Moved ScanResult into Adapters namespace. Pull ScanResult subclasses out of adapter classes (i.e. not inner classes)
