+++ lib/ddr/antivirus/adapters/clamd_scanner_adapter.rb
@@ -1,3 +1,6 @@
+require_relative "scanner_adapter"
+require_relative "scan_result"
+
 module Ddr
   module Antivirus
     module Adapters
@@ -17,37 +20,41 @@ module Ddr
           `clamdscan --no-summary #{path}`.strip
         end
 
-        class ClamdScanResult < Ddr::Antivirus::ScanResult
-          
-          def virus_found
-            if m = /: ([^\s]+) FOUND$/.match(raw)
-              m[1]
-            end
-          end
+      end
 
-          def has_virus?
-            raw =~ / FOUND$/
+      #
+      # Result of a scan with the ClamdScannerAdapter
+      #
+      class ClamdScanResult < ScanResult
+        
+        def virus_found
+          if m = /: ([^\s]+) FOUND$/.match(raw)
+            m[1]
           end
+        end
 
-          def error?
-            raw =~ / ERROR$/
-          end
+        def has_virus?
+          raw =~ / FOUND$/
+        end
 
-          def ok?
-            raw =~ / OK$/
-          end
+        def error?
+          raw =~ / ERROR$/
+        end
 
-          def to_s
-            "#{raw} (#{version})"
-          end
+        def ok?
+          raw =~ / OK$/
+        end
 
-          def default_version
-            `sigtool --version`.strip
-          end
+        def to_s
+          "#{raw} (#{version})"
+        end
 
+        def default_version
+          `sigtool --version`.strip
         end
 
       end
+
     end
   end
 end
