+++ lib/ddr/antivirus/adapters/null_scanner_adapter.rb
@@ -1,3 +1,6 @@
+require_relative "scanner_adapter"
+require_relative "scan_result"
+
 module Ddr
   module Antivirus
     module Adapters
@@ -6,13 +9,17 @@ module Ddr
       #
       class NullScannerAdapter < ScannerAdapter
 
-        class NullScanResult < Ddr::Antivirus::ScanResult; end
-
         def scan(path)
           NullScanResult.new("#{path}: NOT SCANNED - using :null scanner adapter.", path)
         end
 
       end
+
+      #
+      # The result of the scan with the NullScannerAdapter.
+      #
+      class NullScanResult < ScanResult; end
+
     end
   end
 end
