+++ spec/unit/scanner_spec.rb
@@ -1,3 +1,5 @@
+require "ddr/antivirus/adapters/scan_result"
+
 module Ddr
   module Antivirus
     RSpec.describe Scanner do
@@ -11,19 +13,19 @@ module Ddr
           end
         end
         describe "when a virus is found" do
-          before { allow_any_instance_of(ScanResult).to receive(:has_virus?) { true } }
+          before { allow_any_instance_of(Ddr::Antivirus::Adapters::ScanResult).to receive(:has_virus?) { true } }
           it "should raise an execption" do
             expect { subject.scan(path) }.to raise_error
           end
         end
         describe "when a virus is not found" do
-          before { allow_any_instance_of(ScanResult).to receive(:has_virus?) { false } }
+          before { allow_any_instance_of(Ddr::Antivirus::Adapters::ScanResult).to receive(:has_virus?) { false } }
           it "should return the scan result" do
-            expect(subject.scan(path)).to be_a(ScanResult)
+            expect(subject.scan(path)).to be_a(Ddr::Antivirus::Adapters::ScanResult)
           end
         end
         describe "when an error occurs in the scanner" do
-          before { allow_any_instance_of(ScanResult).to receive(:error?) { true } }
+          before { allow_any_instance_of(Ddr::Antivirus::Adapters::ScanResult).to receive(:error?) { true } }
           it "should log an error" do
             expect(Ddr::Antivirus.logger).to receive(:error)
             subject.scan(path)
